BUILD: http: make extract_cookie_value() return an int not size_t
It's very annoying that we have to deal with the crappy size_t and with ints
at some places because these ones don't mix well. Patch 6f61b2 changed the
chunk len to int but its size remains size_t and some functions are having
trouble being used by several callers depending on the type of their arguments.

Let's turn extract_cookie_value() to int for now on, and plan a massive cleanup
later to remove all size_t.

+++ include/types/buffers.h
@@ -167,7 +167,7 @@
 /* describes a chunk of string */
 struct chunk {
 	char *str;	/* beginning of the string itself. Might not be 0-terminated */
-	size_t size;	/* total size of the buffer, 0 if the *str is read-only */
+	int size;	/* total size of the buffer, 0 if the *str is read-only */
 	int len;	/* current size of the string from first to last char. <0 = uninit. */
 };
